Renamed components from index.js -----> <ComponentName>.jsx
